// src/payload/payload-types.ts
export interface Category {
  id: string;
  title: string;
  media?: string | Media;
  parent?: string | Category;
  breadcrumbs?: {
    doc?: string | Category;
    url?: string;
    label?: string;
    id?: string;
  }[];
  updatedAt: string;
  createdAt: string;
}
export interface Media {
  id: string;
  alt: string;
  caption?: {
    [k: string]: unknown;
  }[];
  updatedAt: string;
  createdAt: string;
  url?: string;
  filename?: string;
  mimeType?: string;
  filesize?: number;
  width?: number;
  height?: number;
}
// ...
export interface Spirt {
  id: string;
  title: string;
  media?: string | Media;
  updatedAt: string;
  createdAt: string;
}
export interface Tag {
  id: string;
  title: string;
  updatedAt: string;
  createdAt: string;
}
export interface OtherCategory {
  id: string;
  title: string;
  media?: string | Media;
  updatedAt: string;
  createdAt: string;
}